+++ HISTORY.md
@@ -48,3 +48,9 @@
 ## 0.3.1 (2022-02-04)
 
 * Administration section BugFix getting index file.
+
+## 0.4.0 (2022-02-16)
+
+* Add authorization rules to administration section.
+* Authorization rules CRUD.
+* Authorization checking on microservice call.